requirements/base.txt
amqp==5.3.1
asgiref==3.11.1
asttokens==3.0.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
black==26.3.1
boto3==1.42.71
botocore==1.42.71
cbor2==5.8.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
coverage==7.13.5
cron-descriptor==1.4.5
cryptography==46.0.5
daphne==4.2.1
decorator==5.2.1
distlib==0.4.0
Django==5.2.12
django-anymail==14.0
django-autoslug==1.9.9
django-axes==8.3.1
django-celery-beat==2.9.0
django-cors-headers==4.9.0
django-debug-toolbar==6.2.0
django-environ==0.13.0
django-extensions==4.1
django-filter==25.2
django-ratelimit==4.1.0
django-storages==1.14.6
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
executing==2.2.1
factory_boy==3.3.3
Faker==40.11.0
filelock==3.25.2
flake8==7.3.0
flower==2.0.1
humanize==4.15.0
hyperlink==21.0.0
identify==2.6.18
idna==3.11
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
isort==8.0.1
jedi==0.19.2
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
matplotlib-inline==0.2.1
mccabe==0.7.0
msgpack==1.1.2
mypy_extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
parso==0.8.6
pathspec==1.0.4
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
pre_commit==4.5.1
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.12.1
pyOpenSSL==26.0.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-asyncio>=0.23
pytest-django==4.12.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-discovery==1.2.0
pytokens==0.4.1
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.16.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
stack-data==0.6.3
stripe==14.4.1
tornado==6.5.5
traitlets==5.14.3
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
virtualenv==21.2.0
wcwidth==0.6.0
zope.interface==8.2
zxcvbn==4.5.0
